% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiGam.R
\name{BiGam_CE}
\alias{BiGam_CE}
\title{Get closed-form estimator for Bivariate gamma}
\usage{
BiGam_CE(pars, dat, type = "MLECE", log = TRUE)
}
\arguments{
\item{pars}{parameters of bivariate gamma (alpha1, alpha2, beta).}

\item{dat}{data of bivariate gamma}

\item{type}{output type (MLECE, hessian, score). Default is MLECE}

\item{log}{log-transformation of data. Default is TRUE}
}
\description{
Get closed-form estimator for Bivariate gamma
}
