% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiGam.R
\name{BiGam_MME}
\alias{BiGam_MME}
\title{MME for Bivariate gamma}
\usage{
BiGam_MME(dat, scaletype = "first")
}
\arguments{
\item{dat}{data of bivariate gamma}

\item{scaletype}{scale type for bivariate gamma MME}
}
\description{
MME for Bivariate gamma
}
