% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biwei1.R
\name{BiWei_CE}
\alias{BiWei_CE}
\title{Get root-n consistent closed-form estimator by correlation method and MLE.}
\usage{
BiWei_CE(data)
}
\arguments{
\item{data}{data of bivariate weibull}
}
\description{
Get root-n consistent closed-form estimator by correlation method and MLE.
}
