% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biwei1.R
\name{BiWei_info}
\alias{BiWei_info}
\title{Calculating MLEce for Bivariate weibull}
\usage{
BiWei_info(par_vec, dat, type)
}
\arguments{
\item{par_vec}{parameters of bivariate weibull (alpha1, beta1, alpha2, beta2, delta).}

\item{dat}{data of bivariate weibull}

\item{type}{output type (hessian, MLEce, del, mar)}
}
\description{
Calculating MLEce for Bivariate weibull
}
