% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biwei1.R
\name{BiWei_ML}
\alias{BiWei_ML}
\title{Get MLE for Bivariate weibull}
\usage{
BiWei_ML(data, inits, tol = 1e-07)
}
\arguments{
\item{data}{data of bivariate weibull}

\item{inits}{initial values of iterative algorithm for MLE}

\item{tol}{tolerance for difference.}
}
\description{
Get MLE for Bivariate weibull
}
