% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEce.R
\name{coef.MLEce}
\alias{coef.MLEce}
\alias{print.coef.MLEce}
\title{Extracting estimates.
\code{coef} extracts estimated parameters.}
\usage{
\method{coef}{MLEce}(object, digits = max(3, getOption("digits") - 3), ...)

\method{print}{coef.MLEce}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class "MLEce" made by the function \code{MLEce}.}

\item{digits}{a numeric number of significant digits.}

\item{...}{not used, but exists because of the compatibility.}

\item{x}{an object of class "MLEce".}
}
\value{
estimated parameters are extracted from the "MLEce" class.
}
\description{
Extracting estimates.
\code{coef} extracts estimated parameters.
}
