% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEce.R
\name{computeTime}
\alias{computeTime}
\title{compute MLEce and MLE}
\usage{
computeTime(data, distribution, coef_out = FALSE)
}
\arguments{
\item{data}{a data set.}

\item{distribution}{a character string of a distribution assuming that data set comes from.}

\item{coef_out}{if TRUE, estimated parameters are printed. Default is False.}
}
\value{
a numeric matrix. This matrix include estimated parameters and time.
}
\description{
\code{computeTime} performs a benchmark of MLEce and MLE on a given dataset.
}
\examples{
dat <- rBiWei(n=30, c(4,3,3,4,0.6))
computeTime(dat, "BiWei")

}
