% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biwei1.R
\name{cor_method}
\alias{cor_method}
\title{Get root-n consistent estimator by correlation method.}
\usage{
cor_method(dat)
}
\arguments{
\item{dat}{data of bivariate weibull}
}
\description{
Get root-n consistent estimator by correlation method.
}
