% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiGam.R
\name{dBiGam}
\alias{dBiGam}
\title{log-likelihood for Bivariate gamma}
\usage{
dBiGam(pars, dat1, dat2, log = TRUE)
}
\arguments{
\item{pars}{parameters of bivariate gamma (alpha1, alpha2, beta).}

\item{dat1}{data of marginal column (univariate gamma).}

\item{dat2}{other data of marginal column (univariate gamma).}

\item{log}{log-transformation of data. Default is TRUE}
}
\description{
log-likelihood for Bivariate gamma
}
