% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biwei1.R
\name{dBiWei}
\alias{dBiWei}
\title{Evaluating bivariate Weibull distribution of Gumbel-type}
\usage{
dBiWei(par_vec, dat, log = TRUE)
}
\arguments{
\item{par_vec}{parameters of bivariate weibull (alpha1, beta1, alpha2, beta2, delta).}

\item{dat}{data of bivariate weibull}

\item{log}{log-transformation for data. Default is TRUE.}
}
\description{
Evaluating bivariate Weibull distribution of Gumbel-type
}
