% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biwei1.R
\name{delta_score_probit}
\alias{delta_score_probit}
\title{Calculating delta score}
\usage{
delta_score_probit(par_vec, dat)
}
\arguments{
\item{par_vec}{parameters of bivariate weibull (alpha1, beta1, alpha2, beta2, delta).}

\item{dat}{data of bivariate weibull}
}
\description{
Calculating delta score
}
