% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet.R
\name{Diri_CE_bt}
\alias{Diri_CE_bt}
\title{beta tilde: root-n consistent estimator}
\usage{
Diri_CE_bt(x)
}
\arguments{
\item{x}{data for estimating MLEce}
}
\description{
beta tilde: root-n consistent estimator
}
