% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet.R
\name{Diri_MLE}
\alias{Diri_MLE}
\title{Get MLE for dirichlet}
\usage{
Diri_MLE(x, eps = 1e-10, mxit = 1e+05)
}
\arguments{
\item{x}{data for estimating MLEce}

\item{eps}{epsilon for iterative algorithm.}

\item{mxit}{maximum iteration for MLE. Default is 1e5}
}
\description{
Get MLE for dirichlet
}
