% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEce.R
\docType{data}
\name{flood}
\alias{flood}
\title{The flood events data of the Madawaska basin.}
\format{
A dataframe with 2 variables and 77 observations as follows:
\describe{
\item{Vnorm}{daily average flood volume}
\item{Q}{ flood peak}
}
}
\usage{
data(flood, package = "MLEce")
}
\description{
The data is a subset of the flood events data of the Madawaska basin which is located in the province of Qebec, Canada. Daily streamflow data from 1919 to 1995 are available from HYDAT CD (1998) and Yue (2001).
}
\references{
Environment Canada (1998), HYDAT CD-ROM Version 98-1.05.8: Surface water and sediment data.

Yue. S. (2001) A bivariate gamma distribution for use in multivariate flood frequency analysis. \emph{Hydrological Processes}, 15, 1033–1045.
}
\keyword{datasets}
