% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEce.R
\docType{data}
\name{fossil_pollen}
\alias{fossil_pollen}
\title{The counts data of the frequency of occurrence of different kinds of pollen grains.}
\format{
A 4 variables data frame with 73 observations.
}
\usage{
data(fossil_pollen, package = "MLEce")
}
\description{
It represents the counts data of the frequency of occurrence of different kinds of pollen grains. (Mosimann, 1962).
}
\references{
Mosimann, James E. (1962) \url{doi:10.1093/biomet/49.1-2.65}.
}
\keyword{datasets}
