% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEce.R
\name{gof}
\alias{gof}
\alias{print.gof}
\title{test goodness of fit}
\usage{
gof(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{gof}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object of class "MLEce" made by the function \code{MLEce}.}

\item{digits}{a numeric number of significant digits.}

\item{...}{not used, but exists because of the compatibility.}
}
\value{
\code{gof} returns the p-value of the GCVM test.
}
\description{
test goodness of fit
}
\details{
The null hypothesis of the GCVM test is that "data follows the Bivariate Weibull distribution".
}
\examples{
datt = rBiGam(100, c(1,4,5))
res = MLEce(datt, "BiGam", boots = 50)
gof(res)
}
