% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet.R
\name{llk}
\alias{llk}
\title{Negative Log likelihood function for dirichlet data}
\usage{
llk(x, alp)
}
\arguments{
\item{x}{data for loglikelihood}

\item{alp}{estimated values for the dirichlet parameters}
}
\description{
Negative Log likelihood function for dirichlet data
}
