% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEce.R
\name{MLE_est}
\alias{MLE_est}
\title{Statistical inference of MLE}
\usage{
MLE_est(data, distname, inits, boots = 1000, CI.alpha = 0.05)
}
\arguments{
\item{data}{a numeric vector or matrix.}

\item{distname}{a character string of a distribution assuming that data set comes from (Currently, only Biweibull distribution can be input)}

\item{inits}{a initial vector for MLE.}

\item{boots}{a number of iteration for parametric bootstrapping compute confidence intervals.}

\item{CI.alpha}{a significance level of confidence intervals. default is 0.05.}
}
\value{
\code{MLE_est} returns a list include maximum likelihood estimators and confidence interval of estimated parameters.
}
\description{
Provide value of maximum likelihood estimator, a result of GOF test, and information on the CI of MLE.
}
