% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEce.R
\name{MLEce_est}
\alias{MLEce_est}
\title{Estimate MLEce}
\usage{
MLEce_est(data, distname, boots, CI.alpha)
}
\arguments{
\item{data}{a numeric vector or matrix.}

\item{distname}{a character string "name" naming a distribution for which the corresponding density function dname
and the corresponding distribution function pname must be classically defined.}

\item{boots}{a number of iteration for calculating CI-parametric intervals.}

\item{CI.alpha}{a significance level of confidence intervals. default is 0.05.}
}
\description{
Estimate MLEce
}
