% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEce.R
\name{rBiGam}
\alias{rBiGam}
\title{Generating random data for the bivariate gamma distribution with parameters.}
\usage{
rBiGam(n, paras)
}
\arguments{
\item{n}{number of observations.}

\item{paras}{parameters of bivariate gamma distribution (shape1, shape2, scale).}
}
\value{
\code{rBiGam} generates random deviates. The length of generated data is determined by \code{"n"}.
}
\description{
Generating random data for the bivariate gamma distribution with parameters.
}
\details{
Random generation for the bivariate gamma distribution is presented. The specific generation formulas can be found  in Jang, et al. (2020).
}
\examples{
datt = rBiGam(n=50, c(4,3,3))
}
\references{
Jang, Y.-H., Zhao, J., Kim, H.-M., Yu, K., Kwon, S.and Kim, S. (2023) New closed-form efficient estimator for the multivariate gamma distribution. \emph{Statistica Neerlandica}, 1–18.
}
