% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biwei1.R
\name{rBiWei}
\alias{rBiWei}
\title{Generating random number for the bivariate Weibull distribution with (alpha1, beta1, alpha2, beta2, delta).}
\usage{
rBiWei(n, par_vec)
}
\arguments{
\item{n}{number of observations.}

\item{par_vec}{parameters of BiWeibull (alpha1, beta1, alpha2, beta2, delta).}
}
\value{
\code{rBiWei} generates random deviates.
}
\description{
Generating random number for the bivariate Weibull distribution with (alpha1, beta1, alpha2, beta2, delta).
}
\details{
\code{rBiWei} generates random number data for bivariate weibull distribution.
}
\examples{
datt = rBiWei(n=50, c(4,3,3,4,0.6))

}
