% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEce.R
\name{RNCE_est}
\alias{RNCE_est}
\title{Statistical inference of root-n consistent estimator}
\usage{
RNCE_est(data, distname)
}
\arguments{
\item{data}{a numeric vector or matrix.}

\item{distname}{a character string of a distribution assuming that data set}
}
\value{
a numeric vector of estimated parameters.
}
\description{
Provide value of root-n consistent estimator
}
