% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biwei1.R
\name{Rosen}
\alias{Rosen}
\title{Rosen's transformation for GCVM gof test}
\usage{
Rosen(dat, pars)
}
\arguments{
\item{dat}{data of bivariate weibull}

\item{pars}{parameters of bivariate weibull (alpha1, beta1, alpha2, beta2, delta).}
}
\description{
Rosen's transformation for GCVM gof test
}
