% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet.R
\name{SD2fun.diri}
\alias{SD2fun.diri}
\title{SD2 statistics for GCVM gof test}
\usage{
SD2fun.diri(Data, EST)
}
\arguments{
\item{Data}{data for gof test}

\item{EST}{estimates for gof test}
}
\description{
SD2 statistics for GCVM gof test
}
