% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiGam.R
\name{SD2fun.gam}
\alias{SD2fun.gam}
\title{SD2 statistics for GCVM gof test}
\usage{
SD2fun.gam(Data, EST)
}
\arguments{
\item{Data}{data for gof test}

\item{EST}{estimates for gof test}
}
\description{
SD2 statistics for GCVM gof test
}
