% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet.R
\name{stbz}
\alias{stbz}
\title{stabilization}
\usage{
stbz(x, eps = 1e-10)
}
\arguments{
\item{x}{data for stabilization.}

\item{eps}{epsion for stabilization. It is very small number to extremely small data points}
}
\description{
stabilization
}
