% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEce.R
\docType{data}
\name{flood}
\alias{flood}
\title{The flood events data of the Madawaska basin.}
\format{
A 2 variables data frame with 77 observations.
}
\usage{
data(flood, package = "MLEce")
}
\description{
It represents the flood events data of the Madawaska basin from 1911 to 1995. (Yue, 2001).
}
\references{
Yue, S. (2001). \url{doi:10.1002/hyp.259}.
}
\keyword{datasets}
