% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEce.R
\docType{data}
\name{fossil_pollen}
\alias{fossil_pollen}
\title{The counts data of the frequency of occurrence of different kinds of fossil pollen grains.}
\format{
A dataframe with 73 observations and 4 variables: pinus, abies, quercus and alnus pollens.
}
\usage{
data(fossil_pollen, package = "MLEce")
}
\description{
The data is about the counts of the frequency of occurrence of different kinds of fossil pollen grains and is available in Mosimann (1962).
}
\references{
Mosimann, J. E. (1962) On the compound multinomial distribution, the multivariate beta distribution, and correlations among proportions. \emph{Biometrika}, 49, 65-82.
}
\keyword{datasets}
