% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiGam.R
\name{rBiGam}
\alias{rBiGam}
\title{random generation for the Bivariate Gamma distribution with (shape1, shape2, scale).}
\usage{
rBiGam(n, pars)
}
\arguments{
\item{n}{number of observations.}

\item{pars}{parameters of BiWeibull (shape1, shape2, scale).}
}
\value{
\code{rBiGam} generates random deviates.
}
\description{
random generation for the Bivariate Gamma distribution with (shape1, shape2, scale).
}
\details{
These functions implement formulas given in Hyoung-Moon Kim. et al. (2020). (will be revised.)
}
\examples{
datt = rBiGam(n=50, c(4,3,3))

}
